index, chain/ethereum: Move Ethereum integration into separate module
