+++ index.ts
@@ -1,6 +1,9 @@
 // For testing isolated compilation.
 import 'allocator/arena'
 
+// Ethereum support
+export * from './chain/ethereum'
+
 /**
  * Host store interface.
  */
@@ -8,11 +11,6 @@ export declare namespace store {
   function get(entity: string, id: string): Entity | null
   function set(entity: string, id: string, data: Entity): void
   function remove(entity: string, id: string): void
-}
-
-/** Host Ethereum interface */
-declare namespace ethereum {
-  function call(call: SmartContractCall): Array<EthereumValue> | null
 }
 
 /** Host IPFS interface */
@@ -656,323 +654,6 @@ export class BigDecimal {
   }
 }
 
-/** Type hint for Ethereum values. */
-export enum EthereumValueKind {
-  ADDRESS = 0,
-  FIXED_BYTES = 1,
-  BYTES = 2,
-  INT = 3,
-  UINT = 4,
-  BOOL = 5,
-  STRING = 6,
-  FIXED_ARRAY = 7,
-  ARRAY = 8,
-  TUPLE = 9,
-}
-
-/**
- * Pointer type for EthereumValue data.
- *
- * Big enough to fit any pointer or native `this.data`.
- */
-export type EthereumValuePayload = u64
-
-/**
- * A dynamically typed value used when accessing Ethereum data.
- */
-export class EthereumValue {
-  kind: EthereumValueKind
-  data: EthereumValuePayload
-
-  toAddress(): Address {
-    assert(this.kind == EthereumValueKind.ADDRESS, 'EthereumValue is not an address')
-    return changetype<Address>(this.data as u32)
-  }
-
-  toBoolean(): boolean {
-    assert(this.kind == EthereumValueKind.BOOL, 'EthereumValue is not a boolean.')
-    return this.data != 0
-  }
-
-  toBytes(): Bytes {
-    assert(
-      this.kind == EthereumValueKind.FIXED_BYTES || this.kind == EthereumValueKind.BYTES,
-      'EthereumValue is not bytes.',
-    )
-    return changetype<Bytes>(this.data as u32)
-  }
-
-  toI32(): i32 {
-    assert(
-      this.kind == EthereumValueKind.INT || this.kind == EthereumValueKind.UINT,
-      'EthereumValue is not an int or uint.',
-    )
-    let bigInt = changetype<BigInt>(this.data as u32)
-    return bigInt.toI32()
-  }
-
-  toBigInt(): BigInt {
-    assert(
-      this.kind == EthereumValueKind.INT || this.kind == EthereumValueKind.UINT,
-      'EthereumValue is not an int or uint.',
-    )
-    return changetype<BigInt>(this.data as u32)
-  }
-
-  toString(): string {
-    assert(this.kind == EthereumValueKind.STRING, 'EthereumValue is not a string.')
-    return changetype<string>(this.data as u32)
-  }
-
-  toArray(): Array<EthereumValue> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array.',
-    )
-    return changetype<Array<EthereumValue>>(this.data as u32)
-  }
-
-  toTuple(): EthereumTuple {
-    assert(this.kind == EthereumValueKind.TUPLE, 'EthereumValue is not a tuple.')
-    return changetype<EthereumTuple>(this.data as u32)
-  }
-
-  toTupleArray<T extends EthereumTuple>(): Array<T> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<T>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toTuple() as T
-    }
-    return out
-  }
-
-  toBooleanArray(): Array<boolean> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<boolean>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toBoolean()
-    }
-    return out
-  }
-
-  toBytesArray(): Array<Bytes> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<Bytes>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toBytes()
-    }
-    return out
-  }
-
-  toAddressArray(): Array<Address> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<Address>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toAddress()
-    }
-    return out
-  }
-
-  toStringArray(): Array<string> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<string>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toString()
-    }
-    return out
-  }
-
-  toI32Array(): Array<i32> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<i32>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toI32()
-    }
-    return out
-  }
-
-  toBigIntArray(): Array<BigInt> {
-    assert(
-      this.kind == EthereumValueKind.ARRAY || this.kind == EthereumValueKind.FIXED_ARRAY,
-      'EthereumValue is not an array or fixed array.',
-    )
-    let valueArray = this.toArray()
-    let out = new Array<BigInt>(valueArray.length)
-    for (let i: i32 = 0; i < valueArray.length; i++) {
-      out[i] = valueArray[i].toBigInt()
-    }
-    return out
-  }
-
-  static fromAddress(address: Address): EthereumValue {
-    assert(address.length == 20, 'Address must contain exactly 20 bytes')
-
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.ADDRESS
-    token.data = address as u64
-    return token
-  }
-
-  static fromBoolean(b: boolean): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.BOOL
-    token.data = b ? 1 : 0
-    return token
-  }
-
-  static fromBytes(bytes: Bytes): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.BYTES
-    token.data = bytes as u64
-    return token
-  }
-
-  static fromFixedBytes(bytes: Bytes): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.FIXED_BYTES
-    token.data = bytes as u64
-    return token
-  }
-
-  static fromI32(i: i32): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.INT
-    token.data = BigInt.fromI32(i) as u64
-    return token
-  }
-
-  static fromSignedBigInt(i: BigInt): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.INT
-    token.data = i as u64
-    return token
-  }
-
-  static fromUnsignedBigInt(i: BigInt): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.UINT
-    token.data = i as u64
-    return token
-  }
-
-  static fromString(s: string): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.STRING
-    token.data = s as u64
-    return token
-  }
-
-  static fromArray(values: Array<EthereumValue>): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.ARRAY
-    token.data = values as u64
-    return token
-  }
-
-  static fromTuple(values: EthereumTuple): EthereumValue {
-    let token = new EthereumValue()
-    token.kind = EthereumValueKind.TUPLE
-    token.data = values as u64
-    return token
-  }
-
-  static fromTupleArray(values: Array<EthereumTuple>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromTuple(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromBooleanArray(values: Array<boolean>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromBoolean(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromBytesArray(values: Array<Bytes>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromBytes(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromFixedBytesArray(values: Array<Bytes>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromFixedBytes(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromAddressArray(values: Array<Address>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromAddress(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromStringArray(values: Array<string>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromString(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromI32Array(values: Array<i32>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromI32(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromSignedBigIntArray(values: Array<BigInt>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromSignedBigInt(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-
-  static fromUnsignedBigIntArray(values: Array<BigInt>): EthereumValue {
-    let out = new Array<EthereumValue>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      out[i] = EthereumValue.fromUnsignedBigInt(values[i])
-    }
-    return EthereumValue.fromArray(out)
-  }
-}
-
 /**
  * Enum for supported value types.
  */
@@ -1221,15 +902,6 @@ export class Value {
 }
 
 /**
- * Common representation for Ethereum tuples / Solidity structs.
- *
- * This base class stores the tuple/struct values in an array. The Graph CLI
- * code generation then creates subclasses that provide named getters to
- * access the members by name.
- */
-export class EthereumTuple extends Array<EthereumValue> {}
-
-/**
  * Common representation for entity data, storing entity attributes
  * as `string` keys and the attribute values as dynamically-typed
  * `Value` objects.
@@ -1249,128 +921,6 @@ export class Entity extends TypedMap<string, Value> {
       }
     }
     return target
-  }
-}
-
-/**
- * An Ethereum block.
- */
-export class EthereumBlock {
-  hash: Bytes
-  parentHash: Bytes
-  unclesHash: Bytes
-  author: Address
-  stateRoot: Bytes
-  transactionsRoot: Bytes
-  receiptsRoot: Bytes
-  number: BigInt
-  gasUsed: BigInt
-  gasLimit: BigInt
-  timestamp: BigInt
-  difficulty: BigInt
-  totalDifficulty: BigInt
-  size: BigInt | null
-}
-
-/**
- * An Ethereum transaction.
- */
-export class EthereumTransaction {
-  hash: Bytes
-  index: BigInt
-  from: Address
-  to: Address | null
-  value: BigInt
-  gasUsed: BigInt
-  gasPrice: BigInt
-  input: Bytes
-}
-
-/**
- * Common representation for Ethereum smart contract calls.
- */
-export class EthereumCall {
-  to: Address
-  from: Address
-  block: EthereumBlock
-  transaction: EthereumTransaction
-  inputValues: Array<EthereumEventParam>
-  outputValues: Array<EthereumEventParam>
-}
-
-/**
- * Common representation for Ethereum smart contract events.
- */
-export class EthereumEvent {
-  address: Address
-  logIndex: BigInt
-  transactionLogIndex: BigInt
-  logType: string | null
-  block: EthereumBlock
-  transaction: EthereumTransaction
-  parameters: Array<EthereumEventParam>
-}
-
-/**
- * A dynamically-typed Ethereum event parameter.
- */
-export class EthereumEventParam {
-  name: string
-  value: EthereumValue
-}
-
-class SmartContractCall {
-  contractName: string
-  contractAddress: Address
-  functionName: string
-  functionParams: Array<EthereumValue>
-
-  constructor(
-    contractName: string,
-    contractAddress: Address,
-    functionName: string,
-    functionParams: Array<EthereumValue>,
-  ) {
-    this.contractName = contractName
-    this.contractAddress = contractAddress
-    this.functionName = functionName
-    this.functionParams = functionParams
-  }
-}
-
-/**
- * Low-level interaction with Ethereum smart contracts
- */
-export class SmartContract {
-  _name: string
-  _address: Address
-
-  protected constructor(name: string, address: Address) {
-    this._name = name
-    this._address = address
-  }
-
-  call(name: string, params: Array<EthereumValue>): Array<EthereumValue> {
-    let call = new SmartContractCall(this._name, this._address, name, params)
-    let result = ethereum.call(call)
-    assert(
-      result != null,
-      'Call reverted, probably because an `assert` or `require` in the contract failed, ' +
-        'consider using `try_' +
-        name +
-        '` to handle this in the mapping.',
-    )
-    return result as Array<EthereumValue>
-  }
-
-  tryCall(name: string, params: Array<EthereumValue>): CallResult<Array<EthereumValue>> {
-    let call = new SmartContractCall(this._name, this._address, name, params)
-    let result = ethereum.call(call)
-    if (result == null) {
-      return new CallResult()
-    } else {
-      return CallResult.fromValue(result as Array<EthereumValue>)
-    }
   }
 }
 
@@ -1461,38 +1011,10 @@ export class DataSourceTemplate {
 
 // This is used to wrap a generic so that it can be unioned with `null`, working around limitations
 // with primitives.
-class Wrapped<T> {
+export class Wrapped<T> {
   inner: T
 
   constructor(inner: T) {
     this.inner = inner
-  }
-}
-
-export class CallResult<T> {
-  // `null` indicates a reverted call.
-  private _value: Wrapped<T> | null
-
-  constructor() {
-    this._value = null
-  }
-
-  static fromValue<T>(value: T): CallResult<T> {
-    let result = new CallResult<T>()
-    result._value = new Wrapped(value)
-    return result
-  }
-
-  get reverted(): bool {
-    return this._value == null
-  }
-
-  get value(): T {
-    assert(
-      !this.reverted,
-      'accessed value of a reverted call, ' +
-        'please check the `reverted` field before accessing the `value` field',
-    )
-    return (this._value as Wrapped<T>).inner
   }
 }
